+++ index.js
@@ -39,26 +39,42 @@ NukiBridgePlatform.prototype = {
 
 function NukiLockAccessory(log, config, nukiBridge) {
     this.log = log;
+    this.id = config["id"];
     this.name = config["name"];
     this.nukiBridge = nukiBridge;
     
-    this.service = new Service.LockMechanism(this.name);
-
-    this.service
+    this.informationService = new Service.AccessoryInformation();
+    this.informationService
+        .setCharacteristic(Characteristic.Manufacturer, "Nuki")
+        .setCharacteristic(Characteristic.Model, "Nuki Lock")
+        .setCharacteristic(Characteristic.SerialNumber, "NukiId "+this.id);
+    
+    this.lockService = new Service.LockMechanism(this.name);
+    this.lockService
         .getCharacteristic(Characteristic.LockCurrentState)
         .on('get', this.getState.bind(this));
-
-    this.service
+    this.lockService
         .getCharacteristic(Characteristic.LockTargetState)
         .on('get', this.getState.bind(this))
         .on('set', this.setState.bind(this));
 
+    this.battservice = new Service.BatteryService(this.name);
+    this.battservice
+        .getCharacteristic(Characteristic.BatteryLevel)
+        .on('get', this.getBattery.bind(this));
+    this.battservice
+        .getCharacteristic(Characteristic.ChargingState)
+        .on('get', this.getCharging.bind(this));
+    this.battservice
+        .getCharacteristic(Characteristic.StatusLowBattery)
+        .on('get', this.getLowBatt.bind(this));
+        
     var webHookCallback = (function(isLocked) {
         var newHomeKitState = isLocked ? Characteristic.LockCurrentState.SECURED : Characteristic.LockCurrentState.UNSECURED;
-        this.service.getCharacteristic(Characteristic.LockCurrentState).setValue(newHomeKitState);
+        this.lockService.getCharacteristic(Characteristic.LockCurrentState).setValue(newHomeKitState);
         this.log("HomeKit state change by webhook complete. New isLocked = '%s'.", isLocked);
     }).bind(this);
-    this.nukiLock = new nuki.NukiLock(this.log, nukiBridge, config["id"], config["lock_action"], config["unlock_action"], webHookCallback);
+    this.nukiLock = new nuki.NukiLock(this.log, nukiBridge, this.id, config["lock_action"], config["unlock_action"], webHookCallback);
 };
 
 NukiLockAccessory.prototype.getState = function(callback) {
@@ -77,12 +93,12 @@ NukiLockAccessory.prototype.setState = function(homeKitState, callback) {
             var newHomeKitState = (homeKitState == Characteristic.LockTargetState.SECURED) ?
                 Characteristic.LockCurrentState.SECURED : Characteristic.LockCurrentState.UNSECURED;
             this.log("HomeKit state change complete.");
-            this.service.setCharacteristic(Characteristic.LockCurrentState, newHomeKitState);
+            this.lockService.setCharacteristic(Characteristic.LockCurrentState, newHomeKitState);
             
             if(!update && this.nukiLock.isDoorLatch()) {
                 setTimeout((function(){
                     this.log("HomeKit change for door latch back to locked state complete.");
-                    this.service.getCharacteristic(Characteristic.LockCurrentState).setValue(Characteristic.LockCurrentState.SECURED);
+                    this.lockService.getCharacteristic(Characteristic.LockCurrentState).setValue(Characteristic.LockCurrentState.SECURED);
                     update = true;  
                 }).bind(this), 1000);
             }
@@ -100,6 +116,27 @@ NukiLockAccessory.prototype.setState = function(homeKitState, callback) {
     }
 };
 
+NukiLockAccessory.prototype.getBattery = function(callback) {
+    callback(null, 100);
+};
+
+NukiLockAccessory.prototype.getCharging = function(callback) {
+    callback(null, Characteristic.ChargingState.NOT_CHARGEABLE);
+};
+
+NukiLockAccessory.prototype.getLowBatt = function(callback) {
+    var getLowBattCallback = (function(err, lowBattery){
+        if(err) {
+            this.log("An error occured retrieving battery status. Reason: %s", err);
+            callback(err);
+        }
+        else {
+            callback(null, lowBattery ? Characteristic.StatusLowBattery.BATTERY_LEVEL_LOW : Characteristic.StatusLowBattery.BATTERY_LEVEL_NORMAL);
+        }
+    }).bind(this);
+    this.nukiLock.getLowBatt(getLowBattCallback);
+};
+
 NukiLockAccessory.prototype.getServices = function() {
-  return [this.service];
+  return [this.lockService, this.informationService];
 };